put formatTime method to helpers
