+++ react/src/components/Timeline/components/TimelineEntry.tsx
@@ -1,4 +1,5 @@
 import { Flex, Text } from "@chakra-ui/react";
+import { formatTime } from "utils/helpers";
 import TimelineBullet from "./TimelineBullet";
 
 export interface ITimelineEntryProps {
@@ -7,17 +8,6 @@ export interface ITimelineEntryProps {
 }
 
 function TimelineEntry({ content, time }: ITimelineEntryProps) {
-  function formatTime(date: Date | string): string {
-    const formattedDate = typeof date === "string" && date !== "" ? new Date(date) : date;
-
-    if (formattedDate instanceof Date) {
-      const hours = formattedDate.getHours().toString().padStart(2, "0");
-      const minutes = formattedDate.getMinutes().toString().padStart(2, "0");
-      return `${hours}:${minutes}`;
-    }
-
-    return "";
-  }
   return (
     <Flex flexDirection="column">
       <Flex>
